requirements.txt
falcon==3.0.1
gunicorn==20.1.0
ipython==8.1.1
PyJWT==2.3.0
pytest==7.0.1
pydantic==1.8.2
classic-http-api~=0.0.2
classic-app-layer~=0.0.1
classic-components~=0.0.1
classic-aspects~=0.0.1
classic-sql-storage==0.0.1
classic-messaging-kombu==0.0.3
classic-messaging==0.0.2
backports.cached-property==1.0.1
python-dotenv==0.20.0
marshmallow==3.15.0
kombu==5.1.0
greenlet==1.1.2
gevent==21.12.0
psycopg2-binary==2.9.3
pip install cryptography
